requirements.txt
manim
pyyaml
python-dotenv
requests
httpx
Pillow
numpy
google-genai
